game/requirements/project.txt
--index-url http://dist.bruno.im/index
Django==1.3
django-floppyforms==0.4.5
django-le-social==0.3
gunicorn==0.12.1
django-sentry==1.6.4
